requirements.txt
# Core frameworks
nextcord==3.1.1
python-dotenv==1.0.1

# Cog frameworks
aiohttp==3.13.0
GitPython==3.1.45